Move database update classes to already existing database.updates package.
