+++ src/main/java/cloudgene/mapred/database/updates/BcryptHashUpdate.java
@@ -9,7 +9,6 @@ import org.springframework.security.crypto.bcrypt.BCrypt;
 import cloudgene.mapred.core.User;
 import cloudgene.mapred.database.dao.UserDao;
 import cloudgene.mapred.database.util.Database;
-import cloudgene.mapred.database.util.IUpdateListener;
 
 public class BcryptHashUpdate implements IUpdateListener {
