+++ src/main/java/cloudgene/mapred/server/Application.java
@@ -19,7 +19,7 @@ import cloudgene.mapred.database.updates.BcryptHashUpdate;
 import cloudgene.mapred.database.util.Database;
 import cloudgene.mapred.database.connector.DatabaseConnector;
 import cloudgene.mapred.database.connector.DatabaseConnectorFactory;
-import cloudgene.mapred.database.util.DatabaseUpdater;
+import cloudgene.mapred.database.updates.DatabaseUpdater;
 import cloudgene.mapred.database.util.Fixtures;
 import cloudgene.mapred.jobs.PersistentWorkflowEngine;
 import cloudgene.mapred.jobs.WorkflowEngine;
